feat(filters): publish filters package
