+++ filter/op_binary.go
