+++ filter/op_integer.go
