+++ filter/op_string.go
