+++ filter/operators.go
