+++ filter/operators_test.go
@@ -1,9 +1,10 @@
 package filter_test
 
 import (
-	internal "github.com/elimity-com/scim/internal/filter"
-	"github.com/elimity-com/scim/schema"
 	"testing"
+
+	scimFilter "github.com/elimity-com/scim/filter"
+	"github.com/elimity-com/scim/schema"
 )
 
 // TestValidatorInvalidResourceTypes contains all the cases where an *errors.ScimError gets returned.
@@ -93,7 +94,7 @@ func TestValidatorInvalidResourceTypes(t *testing.T) {
 		},
 	} {
 		t.Run(test.name, func(t *testing.T) {
-			validator, err := internal.NewValidator(test.filter, schema.Schema{
+			validator, err := scimFilter.NewValidator(test.filter, schema.Schema{
 				Attributes: []schema.CoreAttribute{test.attr},
 			})
 			if err != nil {
