+++ filter/path.go
