+++ internal/patch/patch.go
@@ -3,10 +3,11 @@ package patch
 import (
 	"encoding/json"
 	"fmt"
-	f "github.com/elimity-com/scim/internal/filter"
+	"strings"
+
+	f "github.com/elimity-com/scim/filter"
 	"github.com/elimity-com/scim/schema"
 	"github.com/scim2/filter-parser/v2"
-	"strings"
 )
 
 // Op represents the possible value the operation is to perform.
