+++ internal/patch/update.go
@@ -2,7 +2,8 @@ package patch
 
 import (
 	"fmt"
-	f "github.com/elimity-com/scim/internal/filter"
+
+	f "github.com/elimity-com/scim/filter"
 	"github.com/elimity-com/scim/schema"
 )
