+++ server_test.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/elimity-com/scim"
 	"github.com/elimity-com/scim/errors"
-	internal "github.com/elimity-com/scim/internal/filter"
+	f "github.com/elimity-com/scim/filter"
 	"github.com/elimity-com/scim/optional"
 	"github.com/elimity-com/scim/schema"
 	"github.com/scim2/filter-parser/v2"
@@ -156,7 +156,7 @@ func (h testResourceHandler) GetAll(r *http.Request, params scim.ListRequestPara
 			break
 		}
 
-		validator := internal.NewFilterValidator(params.Filter, h.schema)
+		validator := f.NewFilterValidator(params.Filter, h.schema)
 		if err := validator.PassesFilter(v.attributes); err != nil {
 			continue
 		}
